requirements.txt
pillow
pycryptodomex==3.9.4
events==0.3
pyctr==0.1.0